# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Xedaptot API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://xedaptot.onrender.com/api",
      "description": "Production Server"
    },
    {
      "url": "http://localhost:5000/api",
      "description": "Local Server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Paste Firebase ID Token here"
      }
    },
    "schemas": {
...
      "ApiResponseError": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Error message"
          }
        }
      },
...
      "InspectionConditions": {
        "type": "object",
        "properties": {
          "frame": {
            "type": "string",
            "enum": [
              "EXCELLENT",
              "GOOD",
              "FAIR",
              "POOR"
            ]
          },
          "brake": {
            "type": "string",
            "enum": [
              "EXCELLENT",
              "GOOD",
              "FAIR",
              "POOR"
            ]
          },
          "drivetrain": {
            "type": "string",
            "enum": [
              "EXCELLENT",
              "GOOD",
              "FAIR",
              "POOR"
            ]
          },
          "wheels": {
            "type": "string",
            "enum": [
              "EXCELLENT",
              "GOOD",
              "FAIR",
              "POOR"
            ]
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "frame",
          "brake",
          "drivetrain",
          "wheels"
        ]
      },
      "InspectionReport": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "bicycleId": {
            "type": "string"
          },
          "inspectorId": {
            "type": "string"
          },
          "assignedBy": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "REQUESTED",
              "ASSIGNED",
              "IN_PROGRESS",
              "COMPLETED",
              "REJECTED"
            ]
          },
          "conditions": {
            "$ref": "#/components/schemas/InspectionConditions"
          },
          "overallRating": {
            "type": "number",
            "minimum": 1,
            "maximum": 5
          },
          "isPassed": {
            "type": "boolean"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "notes": {
            "type": "string"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time"
          },
          "submittedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Bicycle": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "64d1b5b5b5b5b5b5b5b5b5b5"
          },
          "title": {
            "type": "string",
            "example": "Giant ATX 2022"
          },
          "description": {
            "type": "string",
            "example": "Like new, barely used"
          },
          "price": {
            "type": "number",
            "example": 7500000
          },
          "originalPrice": {
            "type": "number",
            "example": 12000000
          },
          "condition": {
            "type": "string",
            "enum": [
              "NEW",
              "LIKE_NEW",
              "GOOD",
              "FAIR",
              "POOR"
            ]
          },
          "usageMonths": {
            "type": "number",
            "example": 6
          },
          "viewCount": {
            "type": "number",
            "example": 120
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "APPROVED",
              "SOLD",
              "HIDDEN",
              "REJECTED"
            ]
          },
          "isInspected": {
            "type": "boolean",
            "example": false
          },
          "category": {
            "type": "object",
            "properties": {
              "_id": {
                "type": "string"
              },
              "name": {
                "type": "string"
              }
            }
          },
          "brand": {
            "type": "object",
            "nullable": true,
            "properties": {
              "_id": {
                "type": "string"
              },
              "name": {
                "type": "string"
              }
            }
          },
          "seller": {
            "type": "object",
            "properties": {
              "_id": {
                "type": "string"
              },
              "fullName": {
                "type": "string"
              },
              "avatarUrl": {
                "type": "string"
              },
              "reputationScore": {
                "type": "number"
              }
            }
          },
          "specifications": {
            "type": "object",
            "properties": {
              "yearManufactured": {
                "type": "number"
              },
              "frameSize": {
                "type": "string"
              },
              "frameMaterial": {
                "type": "string"
              },
              "wheelSize": {
                "type": "string"
              },
              "gearCount": {
                "type": "number"
              },
              "brakeType": {
                "type": "string"
              },
              "color": {
                "type": "string"
              },
              "weight": {
                "type": "number"
              }
            }
          },
          "location": {
            "type": "object",
            "properties": {
              "address": {
                "type": "string"
              },
              "city": {
                "type": "string"
              },
              "coordinates": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "Point"
                    ]
                  },
                  "coordinates": {
                    "type": "array",
                    "items": {
                      "type": "number"
                    },
                    "description": "[longitude, latitude]"
                  }
                }
              }
            }
          },
          "images": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "url": {
                  "type": "string"
                },
                "mediaType": {
                  "type": "string",
                  "enum": [
                    "image",
                    "video"
                  ]
                },
                "isPrimary": {
                  "type": "boolean"
                },
                "displayOrder": {
                  "type": "number"
                }
              }
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "BicycleInput": {
        "type": "object",
        "required": [
          "title",
          "price",
          "condition",
          "categoryId"
        ],
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "originalPrice": {
            "type": "number"
          },
          "condition": {
            "type": "string",
            "enum": [
              "NEW",
              "LIKE_NEW",
              "GOOD",
              "FAIR",
              "POOR"
            ]
          },
          "usageMonths": {
            "type": "number"
          },
          "categoryId": {
            "type": "string"
          },
          "brandId": {
            "type": "string"
          },
          "specifications": {
            "type": "object"
          },
          "location": {
            "type": "object"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "url": {
                  "type": "string"
                },
                "mediaType": {
                  "type": "string",
                  "enum": [
                    "image",
                    "video"
                  ]
                },
                "isPrimary": {
                  "type": "boolean"
                },
                "displayOrder": {
                  "type": "number"
                }
              }
            }
          }
        }
      },
      "BicycleUpdate": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "originalPrice": {
            "type": "number"
          },
          "condition": {
            "type": "string",
            "enum": [
              "NEW",
              "LIKE_NEW",
              "GOOD",
              "FAIR",
              "POOR"
            ]
          },
          "usageMonths": {
            "type": "number"
          },
          "categoryId": {
            "type": "string"
          },
          "brandId": {
            "type": "string"
          },
          "specifications": {
            "type": "object"
          },
          "location": {
            "type": "object"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "url": {
                  "type": "string"
                },
                "mediaType": {
                  "type": "string",
                  "enum": [
                    "image",
                    "video"
                  ]
                },
                "isPrimary": {
                  "type": "boolean"
                },
                "displayOrder": {
                  "type": "number"
                }
              }
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/bicycles/my": {
      "get": {
        "tags": [
          "Bicycle"
        ],
        "summary": "Get my bicycles",
        "description": "Returns the authenticated user's own bicycle postings with pagination and optional status filter.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "APPROVED",
                "SOLD",
                "HIDDEN",
                "REJECTED"
              ]
            },
            "description": "Filter by bicycle status"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            },
            "description": "Page number (default: 1)"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10,
              "minimum": 1,
              "maximum": 100
            },
            "description": "Items per page (default: 10, max: 100)"
          },
          {
            "in": "query",
            "name": "sort",
            "schema": {
              "type": "string",
              "default": "-createdAt"
            },
            "description": "Sort field (prefix with - for descending, e.g. -createdAt)"
          }
        ],
        "responses": {
          "200": {
            "description": "List of user's bicycles with pagination",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Bicycle"
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "integer",
                          "example": 1
                        },
                        "limit": {
                          "type": "integer",
                          "example": 10
                        },
                        "total": {
                          "type": "integer",
                          "example": 5
                        },
                        "totalPages": {
                          "type": "integer",
                          "example": 1
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - login required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseError"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseError"
                }
              }
            }
          }
        }
      }
    },
    "/bicycles": {
      "get": {
        "tags": [
          "Bicycle"
        ],
        "summary": "Get all bicycles",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "condition",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "NEW",
                "LIKE_NEW",
                "GOOD",
                "FAIR",
                "POOR"
              ]
            }
          },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "brand",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sellerId",
            "in": "query",
            "description": "Filter bicycles by seller ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "minPrice",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "maxPrice",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "city",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "example": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "example": 10
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "example": "-createdAt"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "count": {
                      "type": "integer"
                    },
                    "total": {
                      "type": "integer"
                    },
                    "page": {
                      "type": "integer"
                    },
                    "totalPages": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Bicycle"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "post": {
        "tags": [
          "Bicycle"
        ],
        "summary": "Create new bicycle",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BicycleInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Bicycle created"
          },
          "400": {
            "description": "Missing required fields"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/bicycles/{id}": {
      "get": {
        "tags": [
          "Bicycle"
        ],
        "summary": "Get bicycle by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Bicycle"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Bicycle not found"
          }
        }
      },
      "put": {
        "tags": [
          "Bicycle"
        ],
        "summary": "Update bicycle by ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BicycleUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bicycle updated"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Bicycle not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Bicycle"
        ],
        "summary": "Delete bicycle by ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Bicycle deleted"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Bicycle not found"
          }
        }
      }
    },
    "/bicycles/{id}/status": {
      "put": {
        "tags": [
          "Bicycle"
        ],
        "summary": "Update bicycle status",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status"
                ],
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "PENDING",
                      "APPROVED",
                      "SOLD",
                      "HIDDEN",
                      "REJECTED"
                    ],
                    "example": "APPROVED"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bicycle status updated successfully"
          },
          "400": {
            "description": "Invalid status"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Not authorized to update this bicycle"
          },
          "404": {
            "description": "Bicycle not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
    "/inspectors/bicycles": {
      "get": {
        "tags": [
          "Inspector"
        ],
        "summary": "Get bicycles for inspection",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "filter",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "assigned",
                "available",
                "all"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of bicycles"
          }
        }
      }
    },
...
    "/bicycles/{id}/inspection-report": {
      "get": {
        "tags": [
          "Bicycle"
        ],
        "summary": "Get inspection report for bicycle (public)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Inspection report",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/InspectionReport"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No inspection report found"
          }
        }
      }
    },
    "/bicycles/{id}/request-inspection": {
      "post": {
        "tags": [
          "Bicycle"
        ],
        "summary": "Request inspection for bicycle (seller)",
        "description": "Seller requests inspection for their PENDING bicycle. Sets inspectionStatus to REQUESTED and notifies admins.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Bicycle ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Inspection requested successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Inspection requested successfully. An admin will assign an inspector shortly."
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "bicycleId": {
                          "type": "string"
                        },
                        "title": {
                          "type": "string"
                        },
                        "inspectionStatus": {
                          "type": "string",
                          "example": "REQUESTED"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bicycle not in PENDING status or inspection already requested"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Not the owner of this bicycle"
          },
          "404": {
            "description": "Bicycle not found"
          }
        }
      }
    },
...
  }
}